feat: #refactor - updated images
