+++ src/styles/Home/ExperienceItem.css
@@ -45,7 +45,9 @@
 .experience img {
   width: 50px;
   height: 50px;
+  flex-shrink: 0;
   border-radius: 8px;
   background-color: white;
   object-fit: cover;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.3);
 }
